Moved RedisConfig class to server_util.py
