+++ merlin/server/server_config.py
@@ -318,64 +318,3 @@ def dump_process_file(data, file_path):
     with open(file_path, "w+") as f:
         yaml.dump(data, f, yaml.Dumper)
     return True
-
-
-class RedisConfig:
-    filename = ""
-    entry_order = []
-    entries = {}
-    comments = {}
-    trailing_comments = ""
-    changed = False
-
-    def __init__(self, filename):
-        self.filename = filename
-        self.changed = False
-        self.parse()
-
-    def parse(self):
-        self.entries = {}
-        self.comments = {}
-        with open(self.filename, "r+") as f:
-            file_contents = f.read()
-            file_lines = file_contents.split("\n")
-            comments = ""
-            for line in file_lines:
-                if len(line) > 0 and line[0] != "#":
-                    line_contents = line.split(maxsplit=1)
-                    if line_contents[0] in self.entries:
-                        sub_split = line_contents[1].split(maxsplit=1)
-                        line_contents[0] += " " + sub_split[0]
-                        line_contents[1] = sub_split[1]
-                    self.entry_order.append(line_contents[0])
-                    self.entries[line_contents[0]] = line_contents[1]
-                    self.comments[line_contents[0]] = comments
-                    comments = ""
-                else:
-                    comments += line + "\n"
-            self.trailing_comments = comments[:-1]
-
-    def write(self):
-        with open(self.filename, "w") as f:
-            for entry in self.entry_order:
-                f.write(self.comments[entry])
-                f.write(f"{entry} {self.entries[entry]}\n")
-            f.write(self.trailing_comments)
-
-    def set_filename(self, filename):
-        self.filename = filename
-
-    def set_config_value(self, key: str, value: str):
-        if key not in self.entries:
-            return False
-        self.entries[key] = value
-        self.changed = True
-        return True
-
-    def get_config_value(self, key: str):
-        if key in self.entries:
-            return self.entries[key]
-        return None
-
-    def changes_made(self):
-        return self.changed
