+++ merlin/server/server_util.py
@@ -24,3 +24,70 @@ if __name__ == "__main__":
 
     for i in ipv4_tests:
         print(i, valid_ipv4(i))
+
+
+class RedisConfig:
+    filename = ""
+    entry_order = []
+    entries = {}
+    comments = {}
+    trailing_comments = ""
+    changed = False
+
+    def __init__(self, filename) -> None:
+        self.filename = filename
+        self.changed = False
+        self.parse()
+
+    def parse(self):
+        self.entries = {}
+        self.comments = {}
+        with open(self.filename, "r+") as f:
+            file_contents = f.read()
+            file_lines = file_contents.split("\n")
+            comments = ""
+            for line in file_lines:
+                if len(line) > 0 and line[0] != "#":
+                    line_contents = line.split(maxsplit=1)
+                    if line_contents[0] in self.entries:
+                        sub_split = line_contents[1].split(maxsplit=1)
+                        line_contents[0] += " " + sub_split[0]
+                        line_contents[1] = sub_split[1]
+                    self.entry_order.append(line_contents[0])
+                    self.entries[line_contents[0]] = line_contents[1]
+                    self.comments[line_contents[0]] = comments
+                    comments = ""
+                else:
+                    comments += line + "\n"
+            self.trailing_comments = comments[:-1]
+
+    def write(self):
+        with open(self.filename, "w") as f:
+            for entry in self.entry_order:
+                f.write(self.comments[entry])
+                f.write(f"{entry} {self.entries[entry]}\n")
+            f.write(self.trailing_comments)
+
+    def set_filename(self, filename):
+        self.filename = filename
+
+    def set_config_value(self, key: str, value: str):
+        if key not in self.entries:
+            return False
+        self.entries[key] = value
+        self.changed = True
+        return True
+
+    def get_config_value(self, key: str):
+        if key in self.entries:
+            return self.entries[key]
+        return None
+
+    def changes_made(self):
+        return self.changed
+
+class RedisUsers:
+    filename = ""
+
+    def __init__(self) -> None:
+        pass
